requirements.txt
async-timeout
dnspython
fun-things
generic-consumer
lazy-main
python-dotenv
simple-chalk
typing-extensions
